Using module_info over :erlang.get_module_info
